+++ cli/src/bin/richat-track.rs
@@ -1,7 +1,7 @@
 use {
     clap::Parser,
     futures::{
-        future::{try_join_all, TryFutureExt},
+        future::try_join_all,
         stream::{BoxStream, StreamExt},
     },
     indicatif::{MultiProgress, ProgressBar, ProgressStyle},
@@ -19,13 +19,16 @@ use {
             SubscribeRequestFilterTransactions, SubscribeUpdateBlockMeta, SubscribeUpdateSlot,
             SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
         },
-        richat::GrpcSubscribeRequest,
+        richat::{GrpcSubscribeRequest, RichatFilter},
     },
     serde::Deserialize,
     solana_sdk::clock::Slot,
     std::{
         collections::{BTreeMap, HashMap},
-        sync::Arc,
+        sync::{
+            atomic::{AtomicU64, Ordering},
+            Arc,
+        },
         time::{Duration, SystemTime},
     },
     tokio::{fs, sync::Mutex},
@@ -37,12 +40,18 @@ struct Args {
     /// Path to config
     #[clap(short, long, default_value_t = String::from("config.json"))]
     config: String,
+
+    /// Show only progress, without events
+    #[clap(long, default_value_t = false)]
+    show_events: bool,
 }
 
 #[derive(Debug, Deserialize)]
 #[serde(deny_unknown_fields)]
 struct Config {
-    sources: HashMap<String, ConfigSource>,
+    accounts: bool,
+    transactions: bool,
+    sources: BTreeMap<String, ConfigSource>,
     tracks: Vec<ConfigTrack>,
 }
 
@@ -61,25 +70,42 @@ enum ConfigSource {
 impl ConfigSource {
     async fn subscribe(
         self,
+        accounts_enabled: bool,
+        transactions_enabled: bool,
+    ) -> anyhow::Result<BoxStream<'static, anyhow::Result<UpdateOneof>>> {
+        match self {
+            Self::RichatPluginAgave(config) => {
+                config
+                    .subscribe(accounts_enabled, transactions_enabled)
+                    .await
+            }
+            Self::RichatGrpc(config) => {
+                config
+                    .subscribe(accounts_enabled, transactions_enabled)
+                    .await
+            }
+            Self::YellowstoneGrpc(config) => {
+                config
+                    .subscribe(accounts_enabled, transactions_enabled)
+                    .await
+            }
+        }
+    }
+
+    async fn run_stream(
+        mut stream: BoxStream<'static, anyhow::Result<UpdateOneof>>,
         storage: Arc<Mutex<TrackStorage>>,
         tracks: Vec<ConfigTrack>,
         name: String,
     ) -> anyhow::Result<()> {
-        let mut stream = match self {
-            Self::RichatPluginAgave(config) => config.subscribe().await,
-            Self::RichatGrpc(config) => config.subscribe().await,
-            Self::YellowstoneGrpc(config) => config.subscribe().await,
-        }?;
-
         loop {
             let update = stream
                 .next()
                 .await
                 .ok_or(anyhow::anyhow!("stream finished"))??;
-
             let ts = SystemTime::now();
-
             let mut storage = storage.lock().await;
+            storage.stats.get(&name).unwrap().pb.inc(1);
 
             for track in tracks.iter() {
                 if let Some(slot) = track.matches(&update) {
@@ -108,20 +134,30 @@ enum ConfigSourceRichatPluginAgave {
 }
 
 impl ConfigSourceRichatPluginAgave {
-    async fn subscribe(self) -> anyhow::Result<BoxStream<'static, anyhow::Result<UpdateOneof>>> {
+    async fn subscribe(
+        self,
+        accounts_enabled: bool,
+        transactions_enabled: bool,
+    ) -> anyhow::Result<BoxStream<'static, anyhow::Result<UpdateOneof>>> {
+        let filter = Some(RichatFilter {
+            disable_accounts: !accounts_enabled,
+            disable_transactions: !transactions_enabled,
+            disable_entries: false,
+        });
+
         let stream = match self {
             Self::Quic(config) => {
-                let stream = config.connect().await?.subscribe(None, None).await?;
+                let stream = config.connect().await?.subscribe(None, filter).await?;
                 stream.into_parsed()
             }
             Self::Tcp(config) => {
-                let stream = config.connect().await?.subscribe(None, None).await?;
+                let stream = config.connect().await?.subscribe(None, filter).await?;
                 stream.into_parsed()
             }
             Self::Grpc(config) => {
                 let request = GrpcSubscribeRequest {
                     replay_from_slot: None,
-                    filter: None,
+                    filter,
                 };
 
                 let stream = config.connect().await?.subscribe_richat(request).await?;
@@ -145,25 +181,45 @@ struct ConfigYellowstoneGrpc {
 }
 
 impl ConfigYellowstoneGrpc {
-    async fn subscribe(self) -> anyhow::Result<BoxStream<'static, anyhow::Result<UpdateOneof>>> {
+    async fn subscribe(
+        self,
+        accounts_enabled: bool,
+        transactions_enabled: bool,
+    ) -> anyhow::Result<BoxStream<'static, anyhow::Result<UpdateOneof>>> {
+        let mut accounts = HashMap::new();
+        if accounts_enabled {
+            accounts.insert(
+                "".to_owned(),
+                SubscribeRequestFilterAccounts {
+                    account: vec![],
+                    owner: vec![],
+                    filters: vec![],
+                    nonempty_txn_signature: None,
+                },
+            );
+        }
+
+        let mut transactions = HashMap::new();
+        if transactions_enabled {
+            transactions.insert(
+                "".to_owned(),
+                SubscribeRequestFilterTransactions {
+                    vote: None,
+                    failed: None,
+                    signature: None,
+                    account_include: vec![],
+                    account_exclude: vec![],
+                    account_required: vec![],
+                },
+            );
+        }
+
         let request = SubscribeRequest {
-            accounts: hashmap! { "".to_owned() => SubscribeRequestFilterAccounts {
-                account: vec![],
-                owner: vec![],
-                filters: vec![],
-                nonempty_txn_signature: None,
-            } },
+            accounts,
             slots: hashmap! { "".to_owned() => SubscribeRequestFilterSlots {
                 filter_by_commitment: None
             } },
-            transactions: hashmap! { "".to_owned() => SubscribeRequestFilterTransactions {
-                vote: None,
-                failed: None,
-                signature: None,
-                account_include: vec![],
-                account_exclude: vec![],
-                account_required: vec![],
-            } },
+            transactions,
             transactions_status: HashMap::new(),
             blocks: HashMap::new(),
             blocks_meta: hashmap! { "".to_owned() => SubscribeRequestFilterBlocksMeta {} },
@@ -174,8 +230,14 @@ impl ConfigYellowstoneGrpc {
             from_slot: None,
         };
 
-        let builder = GrpcClient::build_from_shared(self.endpoint)?;
-        let mut client = builder.connect().await?;
+        let mut client = GrpcClient::build_from_shared(self.endpoint)?
+            .tls_config_native_roots(None)
+            .await?
+            .max_decoding_message_size(1024 * 1024 * 1024) // 1GB
+            .connect_timeout(Duration::from_secs(3))
+            .timeout(Duration::from_secs(3))
+            .connect()
+            .await?;
         let stream = client.subscribe_dragons_mouth_once(request).await?;
         let parsed = stream.into_parsed().map(|message| {
             message?
@@ -237,10 +299,11 @@ struct TrackStorage {
     total: usize,
     stats: BTreeMap<String, TrackStat>,
     pb_multi: MultiProgress,
+    show_events: bool,
 }
 
 impl TrackStorage {
-    fn new(names: impl Iterator<Item = String>) -> anyhow::Result<Self> {
+    fn new(names: impl Iterator<Item = String>, show_events: bool) -> anyhow::Result<Self> {
         let mut total = 0;
         let mut stats = BTreeMap::new();
 
@@ -248,7 +311,7 @@ impl TrackStorage {
         for name in names {
             let pb = pb_multi.add(ProgressBar::no_length());
             pb.set_style(ProgressStyle::with_template(&format!(
-                "{{spinner}} {{msg}} | {name}"
+                "{{spinner}} {{msg}} | {name} (total messages: {{pos}})"
             ))?);
 
             total += 1;
@@ -268,6 +331,7 @@ impl TrackStorage {
             total,
             stats,
             pb_multi,
+            show_events,
         })
     }
 
@@ -284,7 +348,9 @@ impl TrackStorage {
         map.insert(name, ts);
 
         if map.len() == self.total {
-            self.pb_multi.println(track.to_info(slot))?;
+            if self.show_events {
+                self.pb_multi.println(track.to_info(slot))?;
+            }
 
             let (best_name, best_ts) = map
                 .iter()
@@ -293,22 +359,35 @@ impl TrackStorage {
                 .unwrap();
             for (name, ts) in map.iter() {
                 if name == &best_name {
-                    self.pb_multi.println(format!("+000.000000s {name}"))?;
+                    if self.show_events {
+                        self.pb_multi.println(format!("+00.000000s {name}"))?;
+                    }
                     self.stats.get_mut(&best_name).unwrap().win += 1;
                 } else {
                     let elapsed = ts.duration_since(best_ts).unwrap();
-                    self.pb_multi.println(format!(
-                        "+{:03}.{:06}s {name}",
-                        elapsed.as_secs(),
-                        elapsed.subsec_micros()
-                    ))?;
+                    if self.show_events {
+                        self.pb_multi.println(format!(
+                            "+{:02}.{:06}s {name}",
+                            elapsed.as_secs(),
+                            elapsed.subsec_micros()
+                        ))?;
+                    }
                     let entry = self.stats.get_mut(name).unwrap();
                     entry.delay += elapsed;
                     entry.delay_count += 1;
                 }
             }
 
-            self.update_pb();
+            // update messages
+            for stat in self.stats.values() {
+                let avg = stat.delay.checked_div(stat.delay_count).unwrap_or_default();
+                stat.pb.set_message(format!(
+                    "win {:010} | avg delay +{:02}.{:06}s",
+                    stat.win,
+                    avg.as_secs(),
+                    avg.subsec_micros()
+                ));
+            }
         }
 
         Ok(())
@@ -317,32 +396,43 @@ impl TrackStorage {
     fn clear(&mut self, finalized: Slot) {
         self.map.retain(|k, _v| *k >= finalized);
     }
-
-    fn update_pb(&self) {
-        for stat in self.stats.values() {
-            let avg = stat.delay.checked_div(stat.delay_count).unwrap_or_default();
-            stat.pb.set_message(format!(
-                "win {:010} | avg delay +{:03}.{:06}s",
-                stat.win,
-                avg.as_secs(),
-                avg.subsec_micros()
-            ));
-        }
-    }
 }
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
+async fn main2() -> anyhow::Result<()> {
     let args = Args::parse();
     let config = fs::read(&args.config).await?;
     let config: Config = serde_yaml::from_slice(&config)?;
 
-    let storage = TrackStorage::new(config.sources.keys().cloned())?;
+    let storage = TrackStorage::new(config.sources.keys().cloned(), args.show_events)?;
     let storage = Arc::new(Mutex::new(storage));
-    try_join_all(config.sources.into_iter().map(|(name, source)| {
-        tokio::spawn(source.subscribe(Arc::clone(&storage), config.tracks.clone(), name))
-            .map_err(anyhow::Error::new)
-    }))
-    .await
-    .map(|_| ())
+
+    let mut futures = vec![];
+    for (name, source) in config.sources {
+        let stream = source
+            .subscribe(config.accounts, config.transactions)
+            .await?;
+        let locked = storage.lock().await;
+        locked.pb_multi.println(format!("connected to {name}"))?;
+        let jh = tokio::spawn(ConfigSource::run_stream(
+            stream,
+            Arc::clone(&storage),
+            config.tracks.clone(),
+            name,
+        ));
+        futures.push(async move { jh.await? });
+    }
+
+    try_join_all(futures).await.map(|_: Vec<()>| ())
+}
+
+fn main() -> anyhow::Result<()> {
+    tokio::runtime::Builder::new_multi_thread()
+        .thread_name_fn(move || {
+            static ATOMIC_ID: AtomicU64 = AtomicU64::new(0);
+            let id = ATOMIC_ID.fetch_add(1, Ordering::Relaxed);
+            format!("richatTrack{id:02}")
+        })
+        .enable_all()
+        .build()?
+        .block_on(main2())
 }